feat: refactor billing components
