+++ src/components/organisms/settings/organization/billing/index.ts
@@ -1,0 +1,3 @@
+export { OrganizationBilling } from "@components/organisms/settings/organization/billing/organizationBilling";
+export { UsageProgressBar } from "@components/organisms/settings/organization/billing/usageProgressBar";
+export { PlanComparisonTable } from "@components/organisms/settings/organization/billing/planComparisonTable";
