+++ src/constants/lists/index.ts
@@ -7,3 +7,4 @@ export {
 	selectIntegrationGoogle,
 	selectIntegrationAws,
 } from "@constants/lists/connections";
+export { getBillingPlanFeatures, type BillingPlanFeature } from "@constants/lists/billingPlanFeatures.constants";
